[#3383] Extract shared SignupConfigResolution module
De-duplicate resolve_autoverify and domain_signup_config from
Core::Controllers::Base and CreateAccount into a shared module.
Each consumer provides two adapter methods for display_domain
and auth settings access.
